+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Fix join-manager typo in action name
   - Fix delete NuvlaBox cluster on last member leaving
   - Generalize use of Docker Socket for NB pull jobs
+  - Add new legacy bind mount for NuvlaBox Update jobs
 
 ## [2.15.2] - 2021-06-01